Move BERT and ALBERT
