+++ src/transformers/configuration_roberta.py
@@ -15,7 +15,7 @@
 # limitations under the License.
 """ RoBERTa configuration """
 
-from .configuration_bert import BertConfig
+from .models.bert.configuration import BertConfig
 from .utils import logging
