+++ src/transformers/tokenization_dpr.py
@@ -19,7 +19,7 @@ import collections
 from typing import List, Optional, Union
 
 from .file_utils import add_end_docstrings, add_start_docstrings
-from .tokenization_bert import BertTokenizer
+from .models.bert.tokenization_slow import BertTokenizer
 from .tokenization_utils_base import BatchEncoding, TensorType
 from .utils import logging
