+++ src/transformers/tokenization_dpr_fast.py
@@ -19,7 +19,7 @@ import collections
 from typing import List, Optional, Union
 
 from .file_utils import add_end_docstrings, add_start_docstrings
-from .tokenization_bert_fast import BertTokenizerFast
+from .models.bert.tokenization_fast import BertTokenizerFast
 from .tokenization_dpr import DPRContextEncoderTokenizer, DPRQuestionEncoderTokenizer, DPRReaderTokenizer
 from .tokenization_utils_base import BatchEncoding, TensorType
 from .utils import logging
